+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.8] - 
 ### 变更
 - 新增 XOrm/orm_init.go 模块的 Source 配置字段根据环境变量求值的功能
+- 新增 XOrm/model_base.go 模块的 OnQuery 接口用于实现自定义的查询逻辑
 
 ### 修复
 - 修复 XOrm/model_base.go 模块 toInt64 函数的类型断言问题